Added Alert for deleting Sync Logs
Added Alert for deleting Sync Logs

+++ OutlookCalender/OutlookCalender/ViewModels/SyncHistoryViewModel.cs
@@ -21,6 +21,7 @@ namespace OutlookCalender.ViewModels
         public bool ApplyFilterEnabled { get { return _applyFilterEnabled; } private set { SetBackingField(ref _applyFilterEnabled, value); } }
         public bool UndoFilterEnabled { get { return _undoFilterEnabled; } private set { SetBackingField(ref _undoFilterEnabled, value); } }
         public bool DeleteHistoryEnabled { get { return _deleteHistoryEnabled; } private set { SetBackingField(ref _deleteHistoryEnabled, value); } }
+        public string DeleteHistoryButtonText { get { return _deleteHistoryButtonText; } private set { SetBackingField(ref _deleteHistoryButtonText, value); } }
 
         private readonly ObservableCollection<SyncLog> _syncLogsInternal;
         private readonly IRepository _repository;
@@ -29,8 +30,12 @@ namespace OutlookCalender.ViewModels
         private bool _applyFilterEnabled;
         private bool _undoFilterEnabled;
         private bool _deleteHistoryEnabled;
+        private string _deleteHistoryButtonText;
+        private readonly Func<string, Task<bool>> _displayAlert;
+        private const string DeleteHistoryButtonTextStandard = "Delete History";
+        private const string DeleteHistoryButtonTextFiltered = "Delete filtered History";
 
-        public SyncHistoryViewModel(IRepository repository)
+        public SyncHistoryViewModel(IRepository repository, Func<string, Task<bool>> displayAlert)
         {
             _repository = repository;
             StartDate = DateTime.Today;
@@ -42,6 +47,8 @@ namespace OutlookCalender.ViewModels
             _syncLogsInternal = new ObservableCollection<SyncLog>();
             SyncLogs = new ReadOnlyObservableCollection<SyncLog>(_syncLogsInternal);
             DeleteHistoryCommand = new Command(DeleteHistory);
+            _displayAlert = displayAlert;
+            DeleteHistoryButtonText = DeleteHistoryButtonTextStandard;
         }
 
         private void OnFilterDateChanged(DateTime oldValue)
@@ -57,7 +64,7 @@ namespace OutlookCalender.ViewModels
                 Task.Run(async () =>
                 {
                     var syncLogs = await _repository.GetAll<SyncLog>();
-                    UpdateSyncLogCollection(syncLogs);
+                    UpdateSyncLogCollection(syncLogs, false);
                 });
    
             }
@@ -69,24 +76,43 @@ namespace OutlookCalender.ViewModels
             Task.Run(async () =>
             {
                 var syncLogs = await _repository.FindAll<SyncLog>(_ => _.StartDate <= _startDate && _.EndDate >= _endDate);
-                 UpdateSyncLogCollection(syncLogs);
-            }).ContinueWith((_) => UndoFilterEnabled = true);
+                 UpdateSyncLogCollection(syncLogs, true);
+            });
    
         }
         
         private void DeleteHistory()
         {
+            const string filtered = "filtered";
+            const string complete = "complete";
+            var message = $"Do you realy want to delete the {(UndoFilterEnabled ? filtered : complete)} History";
+            Device.BeginInvokeOnMainThread(async () =>
+            {
+                var delete = await _displayAlert(message);
+                if(delete)
+                {
 
+                }
+            });
+            
         }
 
         private void DeleteLog(Guid id)
         {
-            var syncLog = _syncLogsInternal.SingleOrDefault(_ => _.Id == id);
-            _syncLogsInternal.Remove(syncLog);
-            _repository.Delete(syncLog);
+
+            Device.BeginInvokeOnMainThread(async () =>
+            {
+                var syncLog = _syncLogsInternal.SingleOrDefault(_ => _.Id == id);
+                var delete = await _displayAlert($"Do you realy want to delete the Log Entry with sync the range {syncLog.StartDate.ToShortDateString()} - {syncLog.EndDate.ToShortDateString()} that was done on {syncLog.SyncDate.ToShortDateString()}");
+                if(delete)
+                {
+                    _syncLogsInternal.Remove(syncLog);
+                    _repository.Delete(syncLog);
+                }
+            });
         }
 
-        private void UpdateSyncLogCollection(IList<SyncLog> syncLogs)
+        private void UpdateSyncLogCollection(IList<SyncLog> syncLogs, bool undoFilteEnabled)
         {
             Device.BeginInvokeOnMainThread(() =>
             {
@@ -95,6 +121,8 @@ namespace OutlookCalender.ViewModels
                     syncLogs.OrderByDescending(_ => _.SyncDate).ToList().ForEach(s => _syncLogsInternal.Add(s));
                 }
                 DeleteHistoryEnabled = _syncLogsInternal.Any();
+                UndoFilterEnabled = undoFilteEnabled;
+                 DeleteHistoryButtonText = undoFilteEnabled ? DeleteHistoryButtonTextFiltered : DeleteHistoryButtonTextStandard;
             });
         }
 
@@ -104,8 +132,8 @@ namespace OutlookCalender.ViewModels
             Task.Run(async () =>
             {
                 var syncLogs = await _repository.GetAll<SyncLog>();
-                UpdateSyncLogCollection(syncLogs);
-            }).ContinueWith((_) => UndoFilterEnabled = false);
+                UpdateSyncLogCollection(syncLogs, false);
+            });
    
         }
     }
